packages/integrations/deepgram/pyproject.toml
[project]
name = "integration-deepgram"
version = "0.1.0"
description = "Deepgram integration for RAW Platform - real-time speech-to-text"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "raw-core",
    "deepgram-sdk>=3.0.0",
    "pipecat-ai>=0.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/integration_deepgram"]
